+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+# 2.0.0 - 2017-06-23
+
+-   Fixed: glob pattern now accept absolute paths.
+-   Changed: `out-dir` now respect `cwd` option.
+
 # 1.0.1 - 2017-06-06
 
 -   Chore: minimum required `file-type` version is now `^5.0.0`.